requirements.txt
python-dotenv>=0.19.0
google-cloud-logging
SQLAlchemy>=1.4.0
psycopg2-binary>=2.9.0
alembic>=1.7.0
google-cloud-firestore
pydantic>=1.8.0
Flask[async]>=3.0.3
requests
pytest 
pytest-mock
aiohttp
httpx
geoalchemy2>=0.13
shapely